+++ CHANGELOG.md
@@ -28,6 +28,7 @@
   - "Biology/Mammals/Harbour seal areas" from NatureMap
   - "Biology/Mammals/Beluga important areas" from NatureMap
   - "Biology/Mammals/Bowhead whale important areas" from NatureMap
+  - "Biology/Fish/Atlantic salmon in Kapisillit" from NatureMap
 
 # v4.0.0alpha1 (2026-03-04)